+++ examples/03_noprefetch/src/App.tsx
@@ -2,6 +2,8 @@ import React, { Suspense } from 'react';
 
 import { ErrorBoundary } from 'react-hooks-fetch';
 
+import { createFetcher } from 'react-hooks-fetch';
+
 import Item from './Item';
 
 const renderError = (error: Error) => (
@@ -11,14 +13,25 @@ const renderError = (error: Error) => (
   </div>
 );
 
+const fetchFunc = async (userId: string) => (await fetch(`https://reqres.in/api/users/${userId}?delay=3`)).json();
+const fetcher = createFetcher<
+  { data: { first_name: string } },
+  string
+>(fetchFunc);
+const fallbackData = { data: { first_name: '' } };
+const lazy1 = fetcher.lazyFetch(fallbackData);
+const lazy2 = fetcher.lazyFetch(fallbackData);
+const lazy3 = fetcher.lazyFetch(fallbackData);
+
 const App: React.FC = () => (
   <ErrorBoundary fallback={renderError}>
     <Suspense fallback={<span>Loading...</span>}>
-      <Item />
+      <Item suspendable={lazy1} />
       <hr />
-      <Item />
+      <Item suspendable={lazy2} />
+      <hr />
+      <Item initialId="9" suspendable={lazy3} />
     </Suspense>
   </ErrorBoundary>
 );
-
 export default App;